Create two sum o(n)
this can be done in o(nlogn) whish also works on leetcode. Sort the array and then walk through the array and for every element check if there is another element using binary search.o(n) works with the code above. Put elements in a hashmap with the indice and the element. Then for every element find a compliment x-vec[i] in this case as it is the only element that come sum from vec[i] to x. Then find it and add the indice and return the answer. Note: in a hashmap, there are no duplicate keys allowed so we check if there are duplicates in the begging while adding it, so we can return answers with same elements and different indices.